+++ KeyDerivationManager.m
@@ -79,7 +79,7 @@
 }
 
 - (void)updateToolTip {
-	[slider setToolTip:[NSString stringWithFormat:NSLocalizedString(@"PBKDF2 iterations: %d", nil), lastHashIterationCount]];
+	[slider setToolTip:[NSString stringWithFormat:NSLocalizedString(@"PBKDF2 iterations: %ld", nil), (long)lastHashIterationCount]];
 }
 
 - (void)mouseUpForKeyDerivationDelaySlider:(KeyDerivationDelaySlider*)aSlider {
